Moved ViewRecipe from components to pages
