+++ app/imports/ui/layouts/App.jsx
@@ -27,7 +27,7 @@ import EditIngredient from '../pages/EditIngredient';
 import AboutUs from '../pages/AboutUs';
 import Vendor from '../pages/Vendor';
 import ListShops from '../pages/ListShops';
-import ViewRecipe from '../components/ViewRecipe';
+import ViewRecipe from '../pages/ViewRecipe';
 
 /** Top-level layout component for this application. Called in imports/startup/client/startup.jsx. */
 const App = () => {
